requirements.txt
flask
pillow
llama-cpp-python